+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkManager.cs
@@ -15,6 +15,9 @@ namespace Unity.Netcode
     [AddComponentMenu("Netcode/Network Manager", -100)]
     public class NetworkManager : MonoBehaviour, INetworkUpdateSystem
     {
+        // TODO: Deprecate...
+        // The following internal values are not used, but because ILPP makes them public in the assembly, they cannot
+        // be removed thanks to our semver validation.
 #pragma warning disable IDE1006 // disable naming rule violation check
 
         // RuntimeAccessModifiersILPP will make this `public`
@@ -409,6 +412,19 @@ namespace Unity.Netcode
         internal NetworkConnectionManager ConnectionManager = new NetworkConnectionManager();
         internal NetworkMessageManager MessageManager = null;
 
+        internal struct Override<T>
+        {
+            private T m_Value;
+            public bool Overidden { get; private set; }
+            internal T Value
+            {
+                get { return Overidden ? m_Value : default(T); }
+                set { Overidden = true; m_Value = value; }
+            }
+        };
+
+        internal Override<ushort> PortOverride;
+
 #if UNITY_EDITOR
         internal static INetworkManagerHelper NetworkManagerHelper;
 
@@ -491,6 +507,15 @@ namespace Unity.Netcode
                 }
             }
         }
+
+        private void ModeChanged(PlayModeStateChange change)
+        {
+            if (IsListening && change == PlayModeStateChange.ExitingPlayMode)
+            {
+                // Make sure we are not holding onto anything in case domain reload is disabled
+                ShutdownInternal();
+            }
+        }
 #endif
 
         /// <summary>
@@ -539,6 +564,9 @@ namespace Unity.Netcode
             NetworkConfig?.InitializePrefabs();
 
             UnityEngine.SceneManagement.SceneManager.sceneUnloaded += OnSceneUnloaded;
+#if UNITY_EDITOR
+            EditorApplication.playModeStateChanged += ModeChanged;
+#endif
         }
 
         private void OnEnable()
@@ -581,13 +609,46 @@ namespace Unity.Netcode
 
         /// <summary>
         /// Sets the maximum size of a single non-fragmented message (or message batch) passed through the transport.
-        /// This should represent the transport's MTU size, minus any transport-level overhead.
+        /// This should represent the transport's default MTU size, minus any transport-level overhead.
+        /// This value will be used for any remote endpoints that haven't had per-endpoint MTUs set.
+        /// This value is also used as the size of the temporary buffer used when serializing
+        /// a single message (to avoid serializing multiple times when sending to multiple endpoints),
+        /// and thus should be large enough to ensure it can hold each message type.
+        /// This value defaults to 1296.
         /// </summary>
         /// <param name="size"></param>
         public int MaximumTransmissionUnitSize
         {
             set => MessageManager.NonFragmentedMessageMaxSize = value & ~7; // Round down to nearest word aligned size
             get => MessageManager.NonFragmentedMessageMaxSize;
+        }
+
+        /// <summary>
+        /// Set the maximum transmission unit for a specific peer.
+        /// This determines the maximum size of a message batch that can be sent to that client.
+        /// If not set for any given client, <see cref="MaximumTransmissionUnitSize"/> will be used instead.
+        /// </summary>
+        /// <param name="clientId"></param>
+        /// <param name="size"></param>
+        public void SetPeerMTU(ulong clientId, int size)
+        {
+            MessageManager.PeerMTUSizes[clientId] = size;
+        }
+
+        /// <summary>
+        /// Queries the current MTU size for a client.
+        /// If no MTU has been set for that client, will return <see cref="MaximumTransmissionUnitSize"/>
+        /// </summary>
+        /// <param name="clientId"></param>
+        /// <returns></returns>
+        public int GetPeerMTU(ulong clientId)
+        {
+            if (MessageManager.PeerMTUSizes.TryGetValue(clientId, out var ret))
+            {
+                return ret;
+            }
+
+            return MessageManager.NonFragmentedMessageMaxSize;
         }
 
         /// <summary>
@@ -609,6 +670,8 @@ namespace Unity.Netcode
             {
                 return;
             }
+
+            ParseCommandLineOptions();
 
             if (NetworkConfig.NetworkTransport == null)
             {
@@ -1013,6 +1076,9 @@ namespace Unity.Netcode
                 OnServerStopped?.Invoke(ConnectionManager.LocalClient.IsClient);
             }
 
+            // In the event shutdown is invoked within OnClientStopped or OnServerStopped, set it to false again
+            m_ShuttingDown = false;
+
             // Reset the client's roles
             ConnectionManager.LocalClient.SetRole(false, false);
 
@@ -1048,6 +1114,40 @@ namespace Unity.Netcode
             {
                 Singleton = null;
             }
+        }
+
+        // Command line options
+        private const string k_OverridePortArg = "-port";
+
+        private string GetArg(string[] commandLineArgs, string arg)
+        {
+            var argIndex = Array.IndexOf(commandLineArgs, arg);
+            if (argIndex >= 0 && argIndex < commandLineArgs.Length - 1)
+            {
+                return commandLineArgs[argIndex + 1];
+            }
+
+            return null;
+        }
+
+        private void ParseArg<T>(string arg, ref Override<T> value)
+        {
+            if (GetArg(Environment.GetCommandLineArgs(), arg) is string argValue)
+            {
+                value.Value = (T)Convert.ChangeType(argValue, typeof(T));
+            }
+        }
+
+        private void ParseCommandLineOptions()
+        {
+#if UNITY_SERVER && UNITY_DEDICATED_SERVER_ARGUMENTS_PRESENT
+            if ( UnityEngine.DedicatedServer.Arguments.Port != null)
+            {
+                PortOverride.Value = (ushort)UnityEngine.DedicatedServer.Arguments.Port;
+            }
+#else
+            ParseArg(k_OverridePortArg, ref PortOverride);
+#endif
         }
     }
 }